+++ submenu.h
@@ -36,7 +36,7 @@ void subMenu(char* patente)
         strToUpper(opcion);
         while(strlen(opcion)!=1 || opcion[0]<'A' || (opcion[0]>'F' && opcion[0]!='V'))
         {
-            cout << "No te hagas el loco: ";
+            cout << "Opcion incorrecta, intente nuevamente: ";
             sys::getline(opcion, 2);
             strToUpper(opcion);
         }
@@ -46,37 +46,44 @@ void subMenu(char* patente)
         switch(opcion[0])
         {
         case 'A':
-        {
-            subMenuMostrar(patente);
-        }break;
+            {
+                subMenuMostrar(patente);
+            }
+            break;
         case 'B':
-        {
-            subMenuEncenderApagar(seleccionado);
-        }break;
+            {
+                subMenuEncenderApagar(seleccionado);
+            }
+            break;
         case 'C':
-        {
-            subMenuCargarNafta(seleccionado);
-        }break;
+            {
+                subMenuCargarNafta(seleccionado);
+            }
+        break;
         case 'D':
-        {
-            subMenuViaje(seleccionado);
-        }break;
+            {
+                subMenuViaje(seleccionado);
+            }
+        break;
         case 'E':
-        {
-            subMenuModificar(seleccionado);
-        }break;
+            {
+                subMenuModificar(seleccionado);
+            }
+        break;
         case 'F':
-        {
-            if(!subMenuEliminar(patente))
-                break;
-        }
+            {
+                if(!subMenuEliminar(patente))
+                    break;
+            }
         case 'V':
-        {
-            salir=true;
-            continue;
-        }break;
+            {
+                salir=true;
+                continue;
+            }
+        break;
         default:
-            cout << "Algo salio mal" << endl;
+            cout << TEXTO_ERROR << endl;
+
         }//FIN SWITCH
 
         pedirEnter();
@@ -92,16 +99,14 @@ void subMenuMostrar(char* patente)
 
 void subMenuEncenderApagar(Coche c)
 {
-    if(c.getMotor())
-    {
-        c.setMotor(false);
+    c.setMotor(!c.getMotor());
+
+    if(!c.getMotor())
         cout << "Motor apagado" << endl;
-    }
+
     else
-    {
-        c.setMotor(true);
         cout << "Motor encendido" << endl;
-    }
+
     guardarCambios(c);
 }
 
@@ -113,7 +118,7 @@ void subMenuCargarNafta(Coche c)
     while(!validarEntero(cadCantidad) || strlen(cadCantidad)>9)
     {
         if(!validarEntero(cadCantidad))
-            cout << "No te hagas el loco: ";
+            cout << "Ingrese solo numeros enteros: ";
         else
             cout << "Por favor, ingrese un numero de no mas de 9 digitos: ";
         sys::getline(cadCantidad, 11);
@@ -123,11 +128,11 @@ void subMenuCargarNafta(Coche c)
     cout << endl;
     if((cantidad+c.getNafta())>999999999)
     {
-        cout << "El tanque tiene una capacidad maxima de 999.999.999, lo 100to" << endl;
+        cout << "El tanque tiene una capacidad maxima de 999.999.999, lo siento" << endl;
     }
     else
     {
-        c.setNafta(cantidad);
+        c.cargarNafta(cantidad);
         cout << "Nafta restante: " << c.getNafta() << endl;
         guardarCambios(c);
     }
@@ -137,7 +142,7 @@ void subMenuViaje(Coche c)
 {
     if(!c.getMotor())
     {
-        cout << "No encendiste el auto capo" << endl;
+        cout << "El motor no esta encendido" << endl;
         return;
     }
     if(c.getNafta()==0)
@@ -145,15 +150,17 @@ void subMenuViaje(Coche c)
         cout << "Tanque vacio!" << endl;
         return;
     }
+
     char cadConsumo[11];
-    cout << "Cuanto vas a gastar de Gasoil hermano? ";
+    cout << "Cuanto desea gastar de nafta? ";
     sys::getline(cadConsumo, 11);
     while(!validarEntero(cadConsumo) || strlen(cadConsumo)>9)
     {
         if(!validarEntero(cadConsumo))
-            cout << "No te hagas el loco: ";
+            cout << "Ingrese solo numeros enteros: ";
         else
             cout << "Por favor, ingrese un numero de no mas de 9 digitos: ";
+
         sys::getline(cadConsumo, 11);
     }
 
@@ -161,15 +168,11 @@ void subMenuViaje(Coche c)
     cout << endl;
     if((c.getNafta()-consumo)<0)
     {
-        cout << "No llegas ni empujando el auto papu, anda a la Shell o tomate un bondi" << endl;
+        cout << "Nafta insuficiente" << endl;
         return;
     }
-    consumo *= -1;
-    c.setNafta(consumo);
-    if(c.getNafta()==0)
-        cout << "No dejaste ni una gota de combustible, suerte para volver" << endl;
-    else
-        cout << "Te queda " << c.getNafta() << " de nafta capo" << endl;
+    c.gastarNafta(consumo);
+    cout << "Te queda " << c.getNafta() << " de nafta" << endl;
     guardarCambios(c);
 }
 
@@ -180,42 +183,20 @@ void subMenuModificar(Coche c)
 
 bool subMenuEliminar(char* patente)
 {
-    cout << "Esta seguro que desea eliminar este auto?" << endl;
-    cout << endl;
     mostrarCoche(buscarCoche(patente));
-    cout << "Si -> S"<< endl;
-    cout << "No -> N"<< endl << endl;
-    char opcion[2];
-    sys::getline(opcion, 2);
-    char op=opcion[0];
-    while(strlen(opcion)!=1 || (op!='s' && op!='S' && op!='n' && op!='N'))
-    {
-        if(strlen(opcion)!=1)
-            cout << "Ingrese un caracter: ";
-        else
-            cout << "No le he entendido, podria repetirme? ";
-        sys::getline(opcion, 2);
-        op=opcion[0];
-    }
-    switch(op)
-    {
-    case 's':
-    case 'S':
+    if( confirmar("\n\nEsta seguro que desea eliminar este auto?(s/n) ") )
     {
         eliminarCoche(patente);
         cout << endl;
-        cout << "Registro eliminado exitosamente" << endl;
-        return true;
+        if( !existeCoche(patente) )
+        {
+            cout << "Registro eliminado exitosamente" << endl;
+            return true;
+        }
+        else
+            cout << TEXTO_ERROR << endl;
     }
-    break;
-    case 'n':
-    case 'N':
-        return false;
-        break;
-    default:
-        cout << "Algo salio mal";
-        return false;
-    }
+    return false;
 }
 
 #endif // SUBMENU_H_INCLUDED
